backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=0.21.0
httpx>=0.23.0
python-multipart>=0.0.5
langchain>=0.0.200
groq>=0.3.0
tiktoken>=0.4.0
numpy>=1.20.0
pandas>=1.5.0
scikit-learn>=1.0.0
twilio>=7.16.0
email-validator>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
